+++ extended.h
@@ -31,7 +31,7 @@ public:
 	Extended (void);
 	virtual ~Extended();
 
-	static Extended * probe (DPContainer &top_level, DPContainer *parent, long long offset, long long size);
+	static Extended * probe (DPContainer &top_level, DPContainer *parent, long offset, long size);
 
 protected:
