+++ filesystem.h
@@ -38,7 +38,7 @@ public:
 protected:
 
 private:
-	long long ext2_get_usage (void);
+	long ext2_get_usage (void);
 
 };
