+++ lvm_volume.cpp
@@ -41,7 +41,7 @@ LvmVolume::~LvmVolume()
 /**
  * read_data
  */
-int LvmVolume::read_data (long long offset, long long size, unsigned char *buffer)
+int LvmVolume::read_data (long offset, long size, unsigned char *buffer)
 {
 	//log_error ("%s\n", __PRETTY_FUNCTION__);
 	//log_error ("%s - %s\n", name.c_str(), device.c_str());
