+++ table.h
@@ -34,7 +34,7 @@ public:
 	static DPContainer * probe (DPContainer &top_level, DPContainer *parent, unsigned char *buffer, int bufsize);
 
 protected:
-	long long fill_space (void);
+	long fill_space (void);
 
 private:
